requirements.txt
addict==2.4.0
ansi2html==1.8.0
anyio==3.3.2
argon2-cffi==21.1.0
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.4
blurhash==1.1.4
boost==0.1
brotlipy==0.7.0
cachetools==4.2.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.5.18.1
cffi==1.14.6
charset-normalizer==2.0.12
clang==5.0
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.10.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7.post1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dpcpp-cpp-rt==2022.2.1
dynamic-reconfigure==1.7.3
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.33.3
gast==0.4.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.15
google-auth-oauthlib==1.0
google-pasta==0.2.0
greenlet==1.1.2
h5py==3.1.0
idna==3.4
image-geometry==1.16.2
imageio==2.9.0
importlib-metadata==4.13.0
importlib-resources==6.0.1
iniconfig==1.1.1
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
interactive-markers==1.12.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.6
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-packaging==0.10.6
jupyter-server==1.24.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
laser_geometry==1.6.7
Markdown==3.3.4
MarkupSafe==2.1.3
Mastodon.py==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.3
message-filters==1.16.0
mistune==3.0.1
mkl==2022.2.1
mkl-fft==1.3.1
mkl-service==2.4.0
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==7.8.0
nbformat==5.7.0
nest-asyncio==1.5.7
networkx==2.6.3
notebook==6.4.4
numpy==1.22.4
numpy-quaternion==2022.4.2
oauthlib==3.2.0
open3d==0.17.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.0.0
plyfile==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycparser==2.20
Pygments==2.16.1
PyJWT==1.7.1
pyparsing==2.4.7
pyquaternion==0.9.9
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-magic==0.4.24
python-qt-binding==0.4.4
pytz==2022.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==25.1.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
Quaternion==3.5.2.post4
referencing==0.30.2
requests==2.28.1
requests-oauthlib==1.3.0
requests-unixsocket==0.3.0
resource_retriever==1.12.7
retrying==1.3.4
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.10.2
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.7.2
rviz==1.14.20
scikit-image==0.18.3
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sensor-msgs==1.13.1
six==1.15.0
sklearn==0.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.2.0
soupsieve==2.5
SQLAlchemy==1.4.25
stack-data==0.6.2
tbb==2021.7.1
tenacity==8.2.3
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
testresources==2.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==2.2.0
tifffile==2021.8.30
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.12.1
topic-tools==1.16.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
tweepy==4.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.12.1
xacro==1.14.16
zipp==3.11.0
